added post

+++ controllers/api/comment-routes.js
@@ -1,54 +1,50 @@
 const router = require('express').Router();
 const { Comment } = require('../../models');
-const Commment = require('../../models/Comment');
 
 router.get('/', (req, res) => {
-Comment.findAll({
-    attributes: [
-        'id',
-        'comment_text',
-        'user_id',
-        'post_id',
-        'created_at',
-      ]
-      .then(dbCommentData => res.json(dbCommentData))
-      .catch(err => {
-          console.log(err);
-          res.status(500).json(err);
-      })
-})
+  Comment.findAll()
+    .then(dbCommentData => res.json(dbCommentData))
+    .catch(err => {
+      console.log(err);
+      res.status(500).json(err);
+    });
 });
 
 router.post('/', (req, res) => {
-Commment.create({
-    comment_text: req.body.comment_text,
-    user_id: req.body.user_id,
-    post_id: req.body.post_id
-})
-.then(dbCommentData => res.json(dbCommentData))
-.catch(err => {
-    console.log(err);
-    res.status(400).json(err);
-});
+  // expects => {comment_text: "This is the comment", user_id: 1, post_id: 2}
+  if (req.session) {
+    Comment.create({
+      comment_text: req.body.comment_text,
+      user_id: req.session.user_id,
+      post_id: req.body.post_id
+    })
+      .then(dbCommentData => res.json(dbCommentData))
+      .catch(err => {
+        console.log(err);
+        res.status(400).json(err);
+      });
+  }
 });
 
 router.delete('/:id', (req, res) => {
-Comment.destroy({
-    where: {
+  if (req.session) {
+    Comment.destroy({
+      where: {
         id: req.params.id
-    }
-    .then(dbCommentData => {
+      }
+    })
+      .then(dbCommentData => {
         if (!dbCommentData) {
-            res.status(404).json({ message: 'No comment found with this id!' });
-            return;
+          res.status(404).json({ message: 'No comment found with this id!' });
+          return;
         }
         res.json(dbCommentData);
-    })
-    .catch(err => {
+      })
+      .catch(err => {
         console.log(err);
         res.status(500).json(err);
-    })
-})
+      });
+  }
 });
 
 module.exports = router;
